+++ tests/example/__init__.py
@@ -1,14 +1,14 @@
 from openapi.middleware import json404
-from openapi.ws import Sockets
+from openapi.ws import Sockets, LocalBroker
 
 from .db import meta
 from .endpoints import routes
-from .ws import ws_routes, LocalPubSub
+from .ws import ws_routes
 
 
 def setup_app(app):
     meta(app['db'].metadata)
-    app['broker'] = LocalPubSub()
+    app['broker'] = LocalBroker()
     app['web_sockets'] = Sockets(app)
     app.middlewares.append(json404)
     app.router.add_routes(routes)
